+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Updated out-dated requirements (latest versions of Flake8 and pycodestyle 
 are incompatible with each other)
 * Updated README with better explanations of the fields
+* Implemented DatePGPPublicKeyField and DateTimePGPPublicKeyField
 
 ## 2.1.1